Berlin Ampcus/ml-service/requirements.txt
fastapi
uvicorn
transformers
torch
pillow
python-multipart
requests
